Rename Analytics* related types to Telemetry*
