+++ src/Umbraco.Cms.ManagementApi/Controllers/Telemetry/TelemetryControllerBase.cs
@@ -2,12 +2,12 @@
 using NSwag.Annotations;
 using Umbraco.New.Cms.Web.Common.Routing;
 
-namespace Umbraco.Cms.ManagementApi.Controllers.Analytics;
+namespace Umbraco.Cms.ManagementApi.Controllers.Telemetry;
 
 [ApiController]
 [VersionedApiBackOfficeRoute("telemetry")]
 [OpenApiTag("Telemetry")]
 [ApiVersion("1.0")]
-public abstract class AnalyticsControllerBase : ManagementApiControllerBase
+public abstract class TelemetryControllerBase : ManagementApiControllerBase
 {
 }
