+++ src/Umbraco.Cms.ManagementApi/ViewModels/Telemetry/TelemetryLevelViewModel.cs
@@ -1,0 +1,10 @@
+﻿using System.Text.Json.Serialization;
+using Umbraco.Cms.Core.Models;
+
+namespace Umbraco.Cms.ManagementApi.ViewModels.Telemetry;
+
+public class TelemetryLevelViewModel
+{
+    [JsonConverter(typeof(JsonStringEnumConverter))]
+    public TelemetryLevel TelemetryLevel { get; set; }
+}
